+++ coinlib/lib/src/scripts/programs/multisig.dart
@@ -1,8 +1,10 @@
 import 'dart:typed_data';
+import 'package:coinlib/src/common/bytes.dart';
 import 'package:coinlib/src/crypto/ec_public_key.dart';
 import 'package:coinlib/src/scripts/operations.dart';
 import 'package:coinlib/src/scripts/program.dart';
 import 'package:coinlib/src/scripts/script.dart';
+import 'package:collection/collection.dart';
 
 class MultisigProgram implements Program {
 
@@ -13,6 +15,43 @@ class MultisigProgram implements Program {
   final Script script;
   late final int threshold;
   late final List<ECPublicKey> pubkeys;
+
+  /// Creates a multisig script program for a given [threshold] (t-of-n) and a
+  /// list of public keys. The public keys are inserted into the script in the
+  /// same order that they are given.
+  MultisigProgram(this.threshold, Iterable<ECPublicKey> pubkeys)
+    : pubkeys = List.unmodifiable(pubkeys),
+    script = Script([
+      ScriptOp.fromNumber(threshold),
+      ...pubkeys.map((pk) => ScriptPushData(pk.data)),
+      ScriptOp.fromNumber(pubkeys.length),
+      checkmultisig,
+    ]) {
+
+      if (pubkeys.isEmpty || pubkeys.length > maxPubkeys) {
+        throw ArgumentError.value(
+          pubkeys, "pubkeys", "must have length between 1 and $maxPubkeys",
+        );
+      }
+
+      if (threshold < 1 || threshold > pubkeys.length) {
+        throw ArgumentError.value(
+          threshold, "threshold",
+          "must have length between 1 and the number of public keys",
+        );
+      }
+
+    }
+
+  /// Creates a multisig script program for a given [threshold] (t-of-n) and a
+  /// list of public keys that are sorted according to the big-endian encoded
+  /// bytes. Public keys will be inserted into the script from smallest to
+  /// largest encoded data.
+  MultisigProgram.sorted(int threshold, Iterable<ECPublicKey> pubkeys)
+    : this(
+      threshold,
+      pubkeys.sorted((a, b) => compareBytes(a.data, b.data)),
+    );
 
   MultisigProgram.fromScript(this.script) {
 
@@ -61,29 +100,5 @@ class MultisigProgram implements Program {
     : this.fromScript(Script.decompile(compiled));
 
   MultisigProgram.fromAsm(String asm) : this.fromScript(Script.fromAsm(asm));
-
-  MultisigProgram(this.threshold, Iterable<ECPublicKey> pubkeys)
-    : pubkeys = List.unmodifiable(pubkeys),
-    script = Script([
-      ScriptOp.fromNumber(threshold),
-      ...pubkeys.map((pk) => ScriptPushData(pk.data)),
-      ScriptOp.fromNumber(pubkeys.length),
-      checkmultisig,
-    ]) {
-
-      if (pubkeys.isEmpty || pubkeys.length > maxPubkeys) {
-        throw ArgumentError.value(
-          pubkeys, "pubkeys", "must have length between 1 and $maxPubkeys",
-        );
-      }
-
-      if (threshold < 1 || threshold > pubkeys.length) {
-        throw ArgumentError.value(
-          threshold, "threshold",
-          "must have length between 1 and the number of public keys",
-        );
-      }
-
-    }
 
 }