+++ coinlib/test/vectors/hd_keys.dart
@@ -1,8 +1,8 @@
 import 'package:coinlib/coinlib.dart';
 import 'package:test/test.dart';
 
-final privPrefix = NetworkParams.mainnet.privHDPrefix;
-final pubPrefix = NetworkParams.mainnet.pubHDPrefix;
+final privPrefix = Network.mainnet.privHDPrefix;
+final pubPrefix = Network.mainnet.pubHDPrefix;
 
 class HDVector {
   final String? seedHex;